requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# miseq-standalone/MiSeq_Standalone/MiSeq_Standalone/asgi.py: 12
# miseq-standalone/MiSeq_Standalone/MiSeq_Standalone/urls.py: 17,18,19
# miseq-standalone/MiSeq_Standalone/MiSeq_Standalone/wsgi.py: 12
# miseq-standalone/MiSeq_Standalone/crispresso_upload/_tests.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/admin.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/apps.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/migrations/0001_initial.py: 3
# miseq-standalone/MiSeq_Standalone/crispresso_upload/migrations/0002_auto_20200616_1451.py: 4
# miseq-standalone/MiSeq_Standalone/crispresso_upload/migrations/0003_auto_20200626_1317.py: 3
# miseq-standalone/MiSeq_Standalone/crispresso_upload/migrations/0004_auto_20200626_1437.py: 3
# miseq-standalone/MiSeq_Standalone/crispresso_upload/migrations/0005_crispressodata_experiment.py: 3
# miseq-standalone/MiSeq_Standalone/crispresso_upload/models.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/templatetags/standalone_filters.py: 2
# miseq-standalone/MiSeq_Standalone/crispresso_upload/templatetags/standalone_tags.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/tests/tests_views.py: 4
# miseq-standalone/MiSeq_Standalone/crispresso_upload/urls.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/views.py: 8,9
# miseq-standalone/MiSeq_Standalone/manage.py: 10
Django

# miseq-standalone/MiSeq_Standalone/crispresso_upload/serializers.py: 1
# miseq-standalone/MiSeq_Standalone/crispresso_upload/views.py: 11,12
djangorestframework

pymemcache

django-cors-headers

fontawesome-free==5.15
